Fix a few issues with LayoutBuilder and GridLayoutBuilder.
Also, allow children to be added and removed from a ListView.

Removing a child requires the removeItem function to be called rather than
removeChild directly. Because only the index-based overload is currently
exposed to scripting, we must perform a list item index lookup. For improved
efficiency, the removeItem(ListViewItem) overload should also be exposed.